Moved pokemon logic to handler
- Added pokemon handler
- Added pokemon handler interface
- Moved the pokemon logic to the pokemon handler
