+++ PokedexAPI/PokedexAPI/Interfaces/IPokemonHandler.cs
@@ -1,0 +1,9 @@
+﻿using PokedexAPI.Models;
+
+namespace PokedexAPI.Interfaces
+{
+    public interface IPokemonHandler
+    {
+        public Task<Pokemon> GetPokemon(string pokemon);
+    }
+}
