+++ PokedexAPI/PokedexAPI/Program.cs
@@ -1,3 +1,6 @@
+using PokedexAPI.Handlers;
+using PokedexAPI.Interfaces;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -6,6 +9,7 @@ builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+builder.Services.AddScoped<IPokemonHandler, PokemonHandler>();
 
 var app = builder.Build();
